+++ mlir/include/mlir-extensions/Conversion/util_to_llvm.hpp
@@ -25,7 +25,7 @@ class Pass;
 
 namespace imex {
 
-/// Convert operations from the plier_util dialect to the LLVM dialect.
+/// Convert operations from the imex_util dialect to the LLVM dialect.
 ///
 /// TODO: We cannot pass LowerToLLVMOptions directly to the pass as it requires
 /// mlir context which is not yet available at this point, pass creation
