+++ mlir/include/mlir-extensions/Dialect/imex_util/dialect.hpp
@@ -24,14 +24,14 @@
 #include <mlir/Interfaces/SideEffectInterfaces.h>
 #include <mlir/Interfaces/ViewLikeInterface.h>
 
-#include "mlir-extensions/Dialect/plier_util/PlierUtilOpsDialect.h.inc"
-#include "mlir-extensions/Dialect/plier_util/PlierUtilOpsEnums.h.inc"
+#include "mlir-extensions/Dialect/imex_util/PlierUtilOpsDialect.h.inc"
+#include "mlir-extensions/Dialect/imex_util/PlierUtilOpsEnums.h.inc"
 
 #define GET_ATTRDEF_CLASSES
-#include "mlir-extensions/Dialect/plier_util/PlierUtilOpsAttributes.h.inc"
+#include "mlir-extensions/Dialect/imex_util/PlierUtilOpsAttributes.h.inc"
 
 #define GET_OP_CLASSES
-#include "mlir-extensions/Dialect/plier_util/PlierUtilOps.h.inc"
+#include "mlir-extensions/Dialect/imex_util/PlierUtilOps.h.inc"
 
 namespace imex {
 namespace util {
