+++ mlir/lib/Conversion/gpu_runtime_to_llvm.cpp
@@ -15,7 +15,7 @@
 #include "mlir-extensions/Conversion/gpu_runtime_to_llvm.hpp"
 
 #include "mlir-extensions/Dialect/gpu_runtime/IR/gpu_runtime_ops.hpp"
-#include "mlir-extensions/Dialect/plier_util/dialect.hpp"
+#include "mlir-extensions/Dialect/imex_util/dialect.hpp"
 #include "mlir-extensions/Transforms/func_utils.hpp"
 
 #include <mlir/Conversion/AsyncToLLVM/AsyncToLLVM.h>
