+++ mlir/lib/Conversion/gpu_to_gpu_runtime.cpp
@@ -15,7 +15,7 @@
 #include "mlir-extensions/Conversion/gpu_to_gpu_runtime.hpp"
 
 #include "mlir-extensions/Dialect/gpu_runtime/IR/gpu_runtime_ops.hpp"
-#include "mlir-extensions/Dialect/plier_util/dialect.hpp"
+#include "mlir-extensions/Dialect/imex_util/dialect.hpp"
 
 #include <mlir/Analysis/BufferViewFlowAnalysis.h>
 #include <mlir/Conversion/ArithmeticToSPIRV/ArithmeticToSPIRV.h>
