+++ mlir/tools/imex-opt/main.cpp
@@ -17,7 +17,7 @@
 #include <mlir/InitAllPasses.h>
 #include <mlir/Tools/mlir-opt/MlirOptMain.h>
 
-#include "mlir-extensions/Dialect/plier_util/dialect.hpp"
+#include "mlir-extensions/Dialect/imex_util/dialect.hpp"
 
 int main(int argc, char **argv) {
   mlir::registerAllPasses();
