+++ numba_dpcomp/numba_dpcomp/mlir_compiler/lib/pipelines/plier_to_std.cpp
@@ -31,8 +31,8 @@
 #include <mlir/Transforms/GreedyPatternRewriteDriver.h>
 #include <mlir/Transforms/Passes.h>
 
+#include "mlir-extensions/Dialect/imex_util/dialect.hpp"
 #include "mlir-extensions/Dialect/plier/dialect.hpp"
-#include "mlir-extensions/Dialect/plier_util/dialect.hpp"
 
 #include "mlir-extensions/Transforms/call_lowering.hpp"
 #include "mlir-extensions/Transforms/cast_utils.hpp"
