+++ numba_dpcomp/numba_dpcomp/mlir_compiler/lib/pipelines/pre_low_simplifications.cpp
@@ -16,8 +16,8 @@
 
 #include "pipelines/base_pipeline.hpp"
 
+#include "mlir-extensions/Dialect/imex_util/dialect.hpp"
 #include "mlir-extensions/Dialect/plier/dialect.hpp"
-#include "mlir-extensions/Dialect/plier_util/dialect.hpp"
 #include "mlir-extensions/Transforms/type_conversion.hpp"
 #include "mlir-extensions/compiler/pipeline_registry.hpp"
